style #2: Code formatting and package structure reorganization
