--- API-Server/src/main/java/swm/hkcc/LGTM/app/modules/auth/dto/TokenDetailDto.java
@@ -1,12 +1,0 @@
-package swm.hkcc.LGTM.app.modules.auth.dto;
-
-import lombok.*;
-import lombok.experimental.SuperBuilder;
-
-@Data
-@ToString(callSuper=true, includeFieldNames=true)
-@SuperBuilder
-public class TokenDetailDto extends TokenDto {
-    private String accessToken;
-    private String refreshToken;
-}
