+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/auth/dto/oauth/GithubOAuthResponse.java
@@ -1,4 +1,4 @@
-package swm.hkcc.LGTM.app.modules.auth.dto;
+package swm.hkcc.LGTM.app.modules.auth.dto.oauth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
