+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/member/repository/MemberRepository.java
@@ -1,11 +1,12 @@
 package swm.hkcc.LGTM.app.modules.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import swm.hkcc.LGTM.app.modules.member.entity.Member;
+import swm.hkcc.LGTM.app.modules.member.domain.Member;
 
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findOneByNickName(String nickName);
+
     Optional<Member> findOneByGithubId(String githubId);
 }
