+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/member/service/CustomUserDetailsService.java
@@ -6,8 +6,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import swm.hkcc.LGTM.app.modules.member.entity.Member;
-import swm.hkcc.LGTM.app.modules.member.entity.custom.CustomUserDetails;
+import swm.hkcc.LGTM.app.modules.member.domain.Member;
+import swm.hkcc.LGTM.app.modules.member.domain.custom.CustomUserDetails;
 import swm.hkcc.LGTM.app.modules.member.repository.MemberRepository;
 
 @Service
